Hide search results when searchbar's onClear event is triggered
This is a temporary workaround, because the SearchBar component should handle this functionality, and should call onChange when onClear is fired. That will be included in a future version of SearchBar, but in the meantime, it's easy enough to workaround by triggering an update for both events in our example wrapper.