Move build step from postinstall to prepublish, & whitelist dist in npm
The postinstall build step is throwing various errors which prevents the package from being installed in other repos. So to fix this, I've removed it from the postinstall step and instead added a prepublish build step, and set it up so that the dist directory will be included when installing the package from npm (indeed, it'll be the only thing included with the package).
This should also drastically improve installation times for this package ⚡️