Add defaultChecked prop support on switch/radio/checkbox
These components already support the `checked` prop, which sets `defaultChecked` on render. However this is an antipattern and causes the following React warning message:

"Warning: f contains an input of type radio with both checked and defaultChecked props. Input elements must be either controlled or uncontrolled (specify either the checked prop, or the defaultChecked prop, but not both). Decide between using a controlled or uncontrolled input element and remove one of these props. More info: https://fb.me/react-controlled-components"

In order to fix this, I've added specific support for the defaultChecked prop, so you can specifically set it, or checked, or both (but don't do that), or neither.