Fix broken styleguide typography
Since changing from rems to  ems, the headings and other font-sizes in the style guide stopped working. This is due to compounding font-sizes, among other things. I've tried to remove the extra font-sizes, and use a base font-size of 10px like Carbon-UI elements, so that both the components and the style-guide are still able to extend the same base typography classes. Unfortunately this has meant that I needed to add in a handful of small specific font-size overrides rather than letting the base font-size take care of it.