gitignore dist files again, & create npmignore to override it.
Previously the dist files were removed from gitignore in order to avoid a postinstall build step. This helped improve install times, but introduced a prerelease step to build and commit compiled code. This code doesn't really need committing, as it's only needed when publishing to npm. Hence this step removes the compiled code from git, and instead adds it to npm only using npmignore. It also removes source files from npm as they're not necessary to include in the download.