Export transpiled, unbundled lib directory
*TLDR: Fixed a packaging issue with KedroUI, improving DX and reducing served bundle size on a example app by around a third*

The way tje package was being served meant that it bundled all components and dependencies into a single compiled JS file and a corresponding CSS file. This was bad for development (as it was all minified and uglified so you couldn’t inspect anything), and bad for production (because it forces the recipient to download the entire UI library and all its bundled dependencies, even if they only want one component. Now, in addition to the bundled minified version, it also serves (as the new default endpoint) a transpiled but unminified version of the source, so that you can import just the bits you need. AFAIK this is best practice.
I tested it out on a new boilerplate React app, which I set to import this version and include only the Input component. When I ran a new build to generate a static production bundle, the resulting JS and CSS code (which includes the app itself, and React, and things like that), was reduced in size by  33%. I’d estimate that for apps using just a handful of components, the savings on the library itself would be well over half. Plus, it’s much nicer to work with now too.