Replace GSAP animations with CSS (#34)
react-gsap-enhancer was throwing errors and warnings in React v16.9. Furthermore, some of the animations and styles weren't working properly anyway.

I realised that GSAP isn't actually necessary for any of these animations, as you can get away with just using CSS transitions instead. So I've remove the need for it, and deleted the dependencies.