Fix z-index bug on styleguide preview panel (#39)
The Preview panel had `position: relative` applied, which applied z-index stacking context, which messed with some of the default styles in dropdowns, search results, etc.

Rather than fix the styles of the components to fix the styleguide, I'd rather just remove the need to override the positioning, by only applying the position relative to the button's container. This way, it gets positioned where we want, but it doesn't affect any other elements.