Add disabled state to Dropdown component (#79)
We needed to be able to properly disable a dropdown over on Kedro-UI. Ideally every component should this property, but for now I'm just adding it to the dropdown.

Setting the `disabled` prop adds a disabled attribute to the label button, which prevents keyboard usage and indicates the state to screen readers etc. I've also faded it out to 50% as a visual indicator, removed the hover, and added a 'not-allowed' cursor.